pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "step-functions-strands-demo"
version = "0.1.0"
description = "AWS Step Functions + Strands Agents for LLM-backed document analysis"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Winston Brown", email = "winston@winstonbrown.me" }]
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "aws-cdk-lib>=2.150.0",
    "constructs>=10.0.0",
    "boto3>=1.34.0",
    "pydantic>=2.7.0",
    "strands-agents>=1.30.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "moto>=5.0.0",
    "respx>=0.21.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"
